[FIX] Restauração e Correções
Lançamento da restauração de recursos perdidos e correções de funcionamento.
